Improve and refactor where jobs values are set
They were set in get_from_source, but now we are returning them
and setting them at return time. This further fixes #9, because
there were 2 sources of the 500's. The issue this resolves is that
the client was being set as an empty dictionary that still existed
if it failed to get the jobs. Now we are only setting the dict if
if succeeds.

This also fixes #6 by adding expiration requirements.